feat(viz): allow tactile interaction in scene3d
Manage 2 finger pinch for zoom and 3 finger for translation.
Allow to simulate right or middle click with a left click (one finger touch):
it allows to interact with window level with a single finger.
Add the slots in the trackball, negato2D and negato3D interactors.
Simulate CTRL+click with long touch pressed: point release state is emitted after 1s if the finger touch the screen and does not move.